requirements.txt
fastapi
uvicorn
yt-dlp
allin1
openai-whisper
demucs
essentia
torch
pydantic